feat: seperate safe 7579 initializer to module
